:sparkles: (jaewon/feature/login) 앱 시작 시 로그인 여부에 따라 맞는 화면을 보여주도록 구현#73 & 관련 파일 그룹화
